+++ src/Framework/ContentPacks/Model/CustomDropBoxData.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.Xna.Framework;
+
+namespace QuestFramework.Framework.ContentPacks.Model
+{
+    public class CustomDropBoxData
+    {
+        public string Name { get; set; }
+        public string Location { get; set; }
+        public Point Tile { get; set; }
+    }
+}